Remove unused contracts
These contracts where included from another repository as part of the base
ERC4337 implementation. Since in this repository we only provide Account and
AccountFactory, we don't need other parts of the protocol.

This PR addresses review issues:
* HAL-02
* HAL-03
* HAL-04
* HAL-05

Merging it will not require redeployment of code, since the code removed was
not imported from MultiSigSmartAccount of MultiSigSmartAccountFactory